Docs update: Relocate essential NextJS installation instructions for easier access
When installing the appkit package for NextJS, the "Extra Configuration" section is essential to running the local server. If the configuration isn't implemented, you're unable to proceed with any of the "Implementation" section. 

Therefore, it makes much more sense for it be located before the "Implementation" section.

This change would've saved me a good chunk of time as I didn't bother scrolling to the bottom. I didn't scroll to the bottom because I was still stuck on the installation step.